pyproject.toml
[project]
name = "banking-data-platform"
version = "2.0.0"
description = "Production-grade banking data engineering platform — PySpark, Kafka, dbt, Airflow, Delta Lake, Redshift"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Mohd Sohail Khan", email = "sohaildani@github.com" }]

dependencies = [
    # Spark + Delta Lake
    "pyspark==3.5.1",
    "delta-spark==3.1.0",
    # Kafka
    "confluent-kafka==2.3.0",
    "fastavro==1.9.4",
    # dbt
    "dbt-core==1.7.8",
    "dbt-redshift==1.7.4",
    # Airflow
    "apache-airflow[amazon,apache.spark]==2.8.3",
    # AWS
    "boto3==1.34.0",
    "botocore==1.34.0",
    # Data quality
    "great-expectations==0.18.12",
    # Utilities
    "pydantic==2.6.0",
    "structlog==24.1.0",
    "tenacity==8.2.3",
    "pandas==2.2.0",
    "pyarrow==15.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.0",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "black==24.2.0",
    "ruff==0.2.2",
    "mypy==1.8.0",
    "pre-commit==3.6.2",
    "moto[s3,kafka,redshift]==5.0.2",
    "chispa==0.9.2",          # PySpark DataFrame comparison in tests
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=spark --cov-report=term-missing"
